fix: run MLX inference in fp16 end-to-end, matching the PT reference (#4)
Weights were fp16 on disk but activations entered as fp32 (default
mx.random.normal latents, np.float32 preprocessing), and MLX type
promotion ran every kernel in fp32 with per-use fp16->fp32 weight
upcasts (~45% of Stage 1 GPU time — see #2).

- Stage 1: add dtype (default fp16) to ShapePipeline mirroring
  pipelines.py, cast weights via set_dtype, fp16 latents/inputs, cast
  scheduler output back to model dtype like diffusers FlowMatchEuler.
- Conditioner: normalize stats + uncond embedding follow module dtype.
- Stage 2: fp16 latents/inputs/view scales, set_dtype on all modules
  (mirrors torch_dtype=float16 in multiview_utils.py), DINO back to
  fp16 like upstream (the documented '5x divergence' does not
  reproduce: cosine 0.999998, mean_rel 1.7% vs fp32 on real weights).
- UNet: cast timestep embedding to sample dtype (missing diffusers
  parity: t_emb.to(sample.dtype)).
- RoPE: cast rotated q/k back to input dtype like PT apply_rotary_emb.
- RealESRGAN: fp16 like upstream half=True.

Verified via smeltr: top kernels all float16, upcast copies gone; DiT
step -32% on the reduced smoke after the companion mlx-arsenal MoE
dtype fix (dgrauet/mlx-arsenal#45). Closes #2.

+++ hy3dpaint/hunyuanpaintpbr_mlx/dino_mlx.py
@@ -180,7 +180,9 @@ class DINOv2MLX(nn.Module):
             align_corners=False,
         )
         patch_new = pt_new.permute(0, 2, 3, 1).numpy().reshape(1, g_new * g_new, -1)
-        new_pos = mx.array(np.concatenate([cls_np, patch_new], axis=1))
+        new_pos = mx.array(
+            np.concatenate([cls_np, patch_new], axis=1)
+        ).astype(self.pos_embed.dtype)
         cache[num_patches_in] = new_pos
         self._pos_embed_cache = cache
         return new_pos
